@@ -6,10 +6,11 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "react": "^16.12.0",
+    "node-sass": "^7.0.1",
+    "react": "^18.0.0",
     "react-connect-context-hooks": "^1.3.1",
-    "react-dom": "^16.12.0",
-    "react-scripts": "3.4.0"
+    "react-dom": "^18.0.0",
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -33,6 +34,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.0.2",
+    "@types/react-dom": "^18.0.0",
     "acorn": "^7.1.1",
     "minimist": "^1.2.5"
   }
