@@ -5,18 +5,19 @@
   "homepage": "https://edriang.github.io/react-connect-context-hooks",
   "devDependencies": {
     "@types/classnames": "^2.2.9",
+    "@types/react-dom": "^18.0.0",
     "acorn": "^7.1.1",
     "minimist": "^1.2.5",
-    "node-sass": "^4.13.1",
-    "react-scripts": "^3.0.0",
+    "node-sass": "^7.0.1",
+    "react-scripts": "^5.0.0",
     "react-test-renderer": "^16.8.6"
   },
   "dependencies": {
-    "axios": "^0.19.2",
+    "axios": "^0.26.1",
     "classnames": "^2.2.6",
-    "react": "^16.8.6",
+    "react": "^18.0.0",
     "react-connect-context-hooks": "^1.9.1",
-    "react-dom": "^16.8.6",
+    "react-dom": "^18.0.0",
     "reselect": "^4.0.0",
     "todomvc-app-css": "^2.2.0"
   },
