@@ -1,10 +1,11 @@
 import React from 'react';
+import { Context } from 'use-context-selector';
 
 import { getProviderContext } from './createContextProvider';
 import { KeyValue } from './typings';
 
 function createMockProvider(Provider: any, Component: React.ReactNode) {
-    const Context: React.Context<any> = getProviderContext(Provider);
+    const Context: Context<any> = getProviderContext(Provider);
 
     return ({ state = {}, actions = {} }: { state: KeyValue, actions: KeyValue }) => (
         <Context.Provider value={{ state, actions }}>
